Simplify and rename Node SDK LLM callbacks to CopilotRequestHandler
Mirror the .NET simplification + terminology rename in the Node SDK:
consolidate the provider/handler two-layer design into a single
copilotRequestHandler.ts, trim the accepted:false plumbing and the
staged backstop, and rename the public Llm* types to Copilot* (types
carry the prefix; properties/methods stay succinct). The session option
becomes requestHandler?: CopilotRequestHandler.

The WebSocket response bridge starts the 101 upgrade head eagerly
(ctx[kBridge].start()) because the runtime gates the connect on it; a
lazy first-write start deadlocks. Generated RPC/wire types are left
untouched.

Drop the mock unit test and the six fabrication e2e tests (covered by
the handler e2e); keep and rename the handler and session-id e2e tests.
