+++ nodejs/src/index.ts
@@ -28,10 +28,10 @@ export {
     approveAll,
     convertMcpCallToolResult,
     createSessionFsAdapter,
+    CopilotRequestHandler,
     CopilotWebSocketHandler,
-    ForwardingWebSocketHandler,
-    LlmRequestHandler,
-    LlmWebSocketCloseStatus,
+    CopilotWebSocketCloseStatus,
+    ForwardingCopilotWebSocketHandler,
     SYSTEM_MESSAGE_SECTIONS,
 } from "./types.js";
 // Re-export the generated session-event types (every *Event interface and
@@ -129,11 +129,7 @@ export type {
     SessionFsSqliteQueryResult,
     SessionFsSqliteQueryType,
     SessionFsSqliteProvider,
-    LlmInferenceConfig,
-    LlmInferenceRequest,
-    LlmInferenceResponseInit,
-    LlmInferenceResponseSink,
-    LlmRequestContext,
+    CopilotRequestContext,
     SystemMessageAppendConfig,
     SystemMessageConfig,
     SystemMessageCustomizeConfig,
