--- nodejs/src/llmInferenceProvider.ts
@@ -1,437 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *--------------------------------------------------------------------------------------------*/
-
-import type {
-    LlmInferenceHandler,
-    LlmInferenceHeaders,
-    LlmInferenceHttpRequestChunkRequest,
-    LlmInferenceHttpRequestChunkResult,
-    LlmInferenceHttpRequestStartRequest,
-    LlmInferenceHttpRequestStartResult,
-} from "./generated/rpc.js";
-import type { createServerRpc } from "./generated/rpc.js";
-
-type ServerRpc = ReturnType<typeof createServerRpc>;
-
-/**
- * An outbound model-layer HTTP request the runtime is asking the SDK
- * consumer to handle on its behalf.
- *
- * This is a low-level shape: URL / method / headers verbatim, body bytes
- * delivered as an async iterable, response delivered through the
- * {@link LlmInferenceResponseSink}. The runtime does not classify the
- * request (no provider type, endpoint kind, wire API). Consumers that
- * need that information derive it themselves from the URL / headers.
- */
-export interface LlmInferenceRequest {
-    /** Opaque runtime-minted id, stable across the request lifecycle. */
-    requestId: string;
-    /**
-     * Id of the runtime session that triggered this request, when one is
-     * in scope. Absent for out-of-session requests (e.g. startup model
-     * catalog).
-     */
-    sessionId?: string;
-    /** HTTP method (`GET`, `POST`, ...). */
-    method: string;
-    /** Absolute URL. */
-    url: string;
-    /** HTTP request headers, multi-valued. */
-    headers: LlmInferenceHeaders;
-    /**
-     * Transport the runtime would otherwise use for this request.
-     * `"http"` (the default) covers plain HTTP and SSE responses;
-     * `"websocket"` indicates a full-duplex message channel where each
-     * {@link requestBody} chunk is one inbound WebSocket message and each
-     * {@link responseBody} write is one outbound message. Consumers branch
-     * on this to decide whether to service the request with an HTTP client
-     * or a WebSocket client.
-     */
-    transport: "http" | "websocket";
-    /**
-     * Request body bytes, yielded as they arrive from the runtime.
-     * Always iterable; an empty body yields zero chunks before completing.
-     */
-    requestBody: AsyncIterable<Uint8Array>;
-    /**
-     * Aborts when the runtime cancels this in-flight request (e.g. the
-     * agent turn was aborted upstream). Pass it straight to `fetch` /
-     * `HttpClient.SendAsync` / your transport so the upstream call is torn
-     * down too. After it fires, writes to {@link responseBody} are ignored.
-     */
-    signal: AbortSignal;
-    /**
-     * Sink the consumer writes the upstream response into. Call
-     * {@link LlmInferenceResponseSink.start} exactly once before writing
-     * body chunks, then one or more {@link LlmInferenceResponseSink.write}
-     * calls, and finish with {@link LlmInferenceResponseSink.end} or
-     * {@link LlmInferenceResponseSink.error}.
-     */
-    responseBody: LlmInferenceResponseSink;
-}
-
-/** Response head passed to {@link LlmInferenceResponseSink.start}. */
-export interface LlmInferenceResponseInit {
-    status: number;
-    statusText?: string;
-    headers?: LlmInferenceHeaders;
-}
-
-/**
- * Sink the consumer writes the upstream response into. The state machine
- * is strict: `start` once → 0..N `write` → exactly one of `end` or
- * `error`. Calling out of order throws.
- */
-export interface LlmInferenceResponseSink {
-    /** Send the response head (status + headers) back to the runtime. */
-    start(init: LlmInferenceResponseInit): Promise<void>;
-    /**
-     * Send a body chunk. `string` is encoded as UTF-8; `Uint8Array` is sent
-     * as binary (base64 on the wire).
-     */
-    write(data: string | Uint8Array): Promise<void>;
-    /** Mark end-of-stream cleanly. */
-    end(): Promise<void>;
-    /** Mark end-of-stream with a transport-level failure. */
-    error(error: { message: string; code?: string }): Promise<void>;
-}
-
-/**
- * Interface for an LLM inference provider. The SDK consumer implements
- * `onLlmRequest`. The same callback handles both buffered and streaming
- * responses — the consumer just calls `responseBody.write` zero or more
- * times before `end`.
- *
- * Use {@link createLlmInferenceAdapter} to convert an
- * {@link LlmInferenceProvider} into the {@link LlmInferenceHandler} the
- * SDK's RPC layer registers.
- */
-export interface LlmInferenceProvider {
-    /**
-     * Called by the runtime once per outbound LLM HTTP request the
-     * consumer has opted to handle. The consumer is responsible for
-     * eventually calling either `responseBody.end()` or
-     * `responseBody.error(...)`; failing to do so leaks runtime state.
-     * Throwing surfaces a transport-level failure to the runtime
-     * (equivalent to `responseBody.error({ message: err.message })`
-     * provided `start` has not yet been called).
-     */
-    onLlmRequest(request: LlmInferenceRequest): Promise<void> | void;
-}
-
-interface BodyQueueItem {
-    chunk?: Uint8Array;
-    end?: boolean;
-    cancel?: { reason?: string };
-}
-
-interface BodyQueue {
-    push(item: BodyQueueItem): void;
-    iterable: AsyncIterable<Uint8Array>;
-}
-
-function makeBodyQueue(): BodyQueue {
-    const buffer: BodyQueueItem[] = [];
-    let waker: (() => void) | null = null;
-    let done = false;
-    const wake = (): void => {
-        const w = waker;
-        waker = null;
-        w?.();
-    };
-    return {
-        push(item: BodyQueueItem): void {
-            buffer.push(item);
-            wake();
-        },
-        iterable: {
-            [Symbol.asyncIterator](): AsyncIterator<Uint8Array> {
-                return {
-                    async next(): Promise<IteratorResult<Uint8Array>> {
-                        if (done) {
-                            return { value: undefined, done: true };
-                        }
-                        while (buffer.length === 0) {
-                            await new Promise<void>((resolve) => {
-                                waker = resolve;
-                            });
-                        }
-                        const item = buffer.shift()!;
-                        if (item.cancel) {
-                            done = true;
-                            const reason = item.cancel.reason
-                                ? `Request cancelled by runtime: ${item.cancel.reason}`
-                                : "Request cancelled by runtime";
-                            throw new Error(reason);
-                        }
-                        if (item.end) {
-                            done = true;
-                            return { value: undefined, done: true };
-                        }
-                        return { value: item.chunk ?? new Uint8Array(), done: false };
-                    },
-                };
-            },
-        },
-    };
-}
-
-const sharedTextEncoder = new TextEncoder();
-
-function decodeChunkData(data: string, binary: boolean): Uint8Array {
-    if (binary) {
-        return new Uint8Array(Buffer.from(data, "base64"));
-    }
-    return sharedTextEncoder.encode(data);
-}
-
-interface PendingState {
-    queue: BodyQueue;
-    started: boolean;
-    finished: boolean;
-    abort: AbortController;
-    cancelled: boolean;
-}
-
-/**
- * Adapt an {@link LlmInferenceProvider} into the generated
- * {@link LlmInferenceHandler} shape consumed by the SDK's RPC dispatcher.
- *
- * Maintains a per-`requestId` state table: each `httpRequestStart`
- * allocates a body queue + response sink and fires
- * `provider.onLlmRequest` in the background. Subsequent `httpRequestChunk`
- * frames are routed into the queue. The sink translates `start` /
- * `write` / `end` / `error` calls into outbound
- * `serverRpc.llmInference.httpResponseStart` / `httpResponseChunk` calls.
- *
- * The handler returns from `httpRequestStart` immediately (synchronously
- * after registering state) so the runtime's RPC reply is not gated on the
- * consumer's I/O. The actual provider work runs asynchronously.
- */
-export function createLlmInferenceAdapter(
-    provider: LlmInferenceProvider,
-    getServerRpc: () => ServerRpc | undefined
-): LlmInferenceHandler {
-    const pending = new Map<string, PendingState>();
-    // Defense-in-depth backstop: chunks that arrive before their `start`
-    // frame (a reordering the runtime's single ordered dispatch should make
-    // impossible) are staged here keyed by requestId and drained the moment
-    // `httpRequestStart` registers the matching state, so a body byte is
-    // never silently dropped.
-    const staged = new Map<string, LlmInferenceHttpRequestChunkRequest[]>();
-
-    function routeChunk(state: PendingState, params: LlmInferenceHttpRequestChunkRequest): void {
-        if (params.cancel) {
-            state.cancelled = true;
-            state.abort.abort();
-            state.queue.push({ cancel: { reason: params.cancelReason } });
-            return;
-        }
-        if (params.data && params.data.length > 0) {
-            state.queue.push({ chunk: decodeChunkData(params.data, !!params.binary) });
-        }
-        if (params.end) {
-            state.queue.push({ end: true });
-        }
-    }
-
-    function makeSink(requestId: string, state: PendingState): LlmInferenceResponseSink {
-        const rpc = (): ServerRpc => {
-            const r = getServerRpc();
-            if (!r) {
-                throw new Error("LLM inference response sink used after RPC connection closed.");
-            }
-            return r;
-        };
-        // The runtime acknowledges every response frame with `accepted`.
-        // `accepted: false` means it has dropped the request (e.g. it
-        // cancelled), so we abort the provider's upstream work and stop
-        // emitting — there is no consumer for further frames.
-        const rejectedByRuntime = (): never => {
-            if (!state.cancelled) {
-                state.cancelled = true;
-                state.abort.abort();
-            }
-            state.finished = true;
-            pending.delete(requestId);
-            throw new Error(
-                "LLM inference response was rejected by the runtime (request no longer active)."
-            );
-        };
-        return {
-            async start(init: LlmInferenceResponseInit): Promise<void> {
-                if (state.started) {
-                    throw new Error("LLM inference response sink.start() called twice.");
-                }
-                if (state.finished) {
-                    throw new Error("LLM inference response sink already finished.");
-                }
-                state.started = true;
-                const result = await rpc().llmInference.httpResponseStart({
-                    requestId,
-                    status: init.status,
-                    statusText: init.statusText,
-                    headers: init.headers ?? {},
-                });
-                if (!result.accepted) {
-                    rejectedByRuntime();
-                }
-            },
-            async write(data: string | Uint8Array): Promise<void> {
-                if (state.cancelled) {
-                    throw new Error("LLM inference request was cancelled by the runtime.");
-                }
-                if (!state.started) {
-                    throw new Error("LLM inference response sink.write() called before start().");
-                }
-                if (state.finished) {
-                    throw new Error(
-                        "LLM inference response sink.write() called after end()/error()."
-                    );
-                }
-                const isString = typeof data === "string";
-                const result = await rpc().llmInference.httpResponseChunk({
-                    requestId,
-                    data: isString ? data : Buffer.from(data).toString("base64"),
-                    binary: !isString,
-                    end: false,
-                });
-                if (!result.accepted) {
-                    rejectedByRuntime();
-                }
-            },
-            async end(): Promise<void> {
-                if (state.finished) {
-                    return;
-                }
-                state.finished = true;
-                pending.delete(requestId);
-                await rpc().llmInference.httpResponseChunk({
-                    requestId,
-                    data: "",
-                    end: true,
-                });
-            },
-            async error(err: { message: string; code?: string }): Promise<void> {
-                if (state.finished) {
-                    return;
-                }
-                state.finished = true;
-                pending.delete(requestId);
-                await rpc().llmInference.httpResponseChunk({
-                    requestId,
-                    data: "",
-                    end: true,
-                    error: { message: err.message, code: err.code },
-                });
-            },
-        };
-    }
-
-    async function failViaSink(
-        sink: LlmInferenceResponseSink,
-        state: PendingState,
-        message: string
-    ): Promise<void> {
-        if (state.finished) {
-            return;
-        }
-        try {
-            if (!state.started) {
-                await sink.start({ status: 502, headers: {} });
-            }
-            await sink.error({ message });
-        } catch {
-            // Best-effort — the connection may already be dead.
-        }
-    }
-
-    async function finishCancelled(
-        sink: LlmInferenceResponseSink,
-        state: PendingState
-    ): Promise<void> {
-        if (state.finished) {
-            return;
-        }
-        try {
-            if (!state.started) {
-                await sink.start({ status: 499, headers: {} });
-            }
-            await sink.error({ message: "Request cancelled by runtime", code: "cancelled" });
-        } catch {
-            // Best-effort — the runtime already dropped the request on cancel.
-        }
-    }
-
-    return {
-        async httpRequestStart(
-            params: LlmInferenceHttpRequestStartRequest
-        ): Promise<LlmInferenceHttpRequestStartResult> {
-            const state: PendingState = {
-                queue: makeBodyQueue(),
-                started: false,
-                finished: false,
-                abort: new AbortController(),
-                cancelled: false,
-            };
-            pending.set(params.requestId, state);
-            const stagedChunks = staged.get(params.requestId);
-            if (stagedChunks) {
-                staged.delete(params.requestId);
-                for (const chunk of stagedChunks) {
-                    routeChunk(state, chunk);
-                }
-            }
-            const sink = makeSink(params.requestId, state);
-            const request: LlmInferenceRequest = {
-                requestId: params.requestId,
-                sessionId: params.sessionId,
-                method: params.method,
-                url: params.url,
-                headers: params.headers,
-                transport: params.transport ?? "http",
-                requestBody: state.queue.iterable,
-                signal: state.abort.signal,
-                responseBody: sink,
-            };
-            void (async () => {
-                try {
-                    await provider.onLlmRequest(request);
-                    if (!state.finished) {
-                        await failViaSink(
-                            sink,
-                            state,
-                            "LLM inference provider returned without finalising the response (call responseBody.end() or .error())."
-                        );
-                    }
-                } catch (err) {
-                    if (state.cancelled || state.abort.signal.aborted) {
-                        // The runtime already cancelled this request; the
-                        // provider's throw is just the abort propagating
-                        // out of its upstream call. Acknowledge with a
-                        // terminal cancelled error if we still can.
-                        await finishCancelled(sink, state);
-                        return;
-                    }
-                    const message = err instanceof Error ? err.message : String(err);
-                    await failViaSink(sink, state, message);
-                }
-            })();
-            return {};
-        },
-        async httpRequestChunk(
-            params: LlmInferenceHttpRequestChunkRequest
-        ): Promise<LlmInferenceHttpRequestChunkResult> {
-            const state = pending.get(params.requestId);
-            if (!state) {
-                const buffered = staged.get(params.requestId) ?? [];
-                buffered.push(params);
-                staged.set(params.requestId, buffered);
-                return {};
-            }
-            routeChunk(state, params);
-            return {};
-        },
-    };
-}
